Emit classExpression to a raw ClassDeclaration
when --isolatedDeclaration is on.

Emitting a type for a variable that is being assigned to
a class Expression is complex, and as isolatedDeclaration is
purely static, it is difficult to reason for type properly coming
from an extends clause, as it requires further understanding what
its being extended from, so this goes beyond a simple traversal.

Rather, emitting the class expression into class declaration makes
it a bit easier, as we have to worry about the types on the properties
only.

Signed-off-by: Hana Joo <hanajoo@google.com>